requirements.txt
protobuf
cpm_kernels
torch>=2.0
jieba==0.42.1
lxml==4.9.4
gradio
mdtex2html
sentencepiece
sse-starlette
streamlit>=1.24.0
pyfunctional==1.3.0
chardet
faiss-cpu
PyMuPDF
openai
llama-index